Today 90 percent App work done

+++ src/components/App Components/MovieCard.js
@@ -7,15 +7,22 @@ const MovieCard = (props) => {
     const { id, movieName, rank } = props
 
     const handleDelete = (id) => {
-        dispatch(deleteItem(id))
+        const confirm = window.confirm('Are you Sure ?')
+        if (confirm) {
+            dispatch(deleteItem(id))
+        }
     }
 
     return (
         <div className='card'>
-            <div className="card-body">
-                <h5 className="card-title">{movieName}</h5>
-                <p className="card-text"># {rank}</p>
-                <button className='btn btn-danger' onClick={() => handleDelete(id)}>Delete</button>
+            <div className="card-body d-flex">
+                <img src="logo192.png" alt="moviePoster" />
+                <div>
+                    <h5 className="card-title">{movieName}</h5>
+                    <p className="card-text"># {rank}</p>
+                    <button className='btn btn-danger' onClick={() => handleDelete(id)}>Delete</button>
+                </div>
+
             </div>
         </div>
     )
